requirements.txt
python-binance==1.0.29
pandas==2.3.1
ta==0.11.0
requests==2.32.4
numpy==2.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
aiohttp==3.12.14
websockets==15.0.1
